# openapi8-8-22.yaml
openapi: 3.0.2
servers:
  - url: http://localhost:8080
info:
  description: Esta es la API que vamos a crear en clase
  version: "1.0"
  title: Delivery API
...
paths:
...
  /users:
     get:
      tags: 
        - users
      summary: Listado de usuarios
      description: Devuelve el listado completo de todas los usuarios registrados. Es posible filtrar por el apellido del usuario
      operationId: getUsers
      parameters:
        - name: surname
          in: query
          description: apellido del usuario a filtrar
          required: false
          schema:
            type: string
            default: "Lopez"
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado usuario con ese apellido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el usuario"
     post:
      tags:
        - users
      summary: Registrar un users
      description: Registra un nuevo usuario
      operationId: postUser
      requestBody:
        description: Los datos de un nuevo usuario
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIn'
      responses:
        '201':
          description: Se ha registrado el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '400':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
  /user/{userId}:
    get:
      tags: 
        - users
      summary: Obtiene los datos de los usuarios
      description: Devuelve los datos de un usuario determinado
      operationId: getUser
      parameters:
        - name: userId
          in: path
          description: Identificador del usuario
          required: true
          schema:
            type: number
            format: int64
            example: 1
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el usuario"
    delete:
      tags: 
        - users
      summary: Da de baja un usuario
      description: Da de baja un usuario determinado
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: Identificador del usuario
          required: true
          schema:
            type: number
            format: int64
            example: 23
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el usuario"
    put:
      tags: 
        - users
      summary: Modifica un usuario
      description: Modifica los datos de un usuario determinado
      operationId: userRestaurant
      parameters:
        - name: userId
          in: path
          description: Identificador del usuario que se va a modificar
          required: true
          schema:
            type: number
            format: int64
            example: 23
      requestBody:
        description: Los datos nuevos del usuario
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIn'
      responses:
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: No se ha encontrado el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el usuario"
                
                
    patch:
      tags: 
        - users
      summary: Modifica el estado de un usuario
      description: Modifica la direccion del usuario
      operationId: patchUser
      parameters:
        - name: userId
          in: path
          description: string que cambia la direccion del usuario
          required: true
          schema:
            type: string
            example: "calle la paz n8"
      requestBody:
        description: Los datos nuevos del usuario
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIn'
      responses:
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: No se ha encontrado el restaurante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el usuario"
...
components:
  schemas:
...
    UserIn:
      properties:
        dni:
          type: string
          example: "94839042A"
        name:
          type: string
          example: "Aitor"
        surname:
          type: string
          example: "Sanchez"
        birthDate:
          type: string
          format: date
          example: "1998-02-01"
        address:
          type: string
          example: "calle salomon 8-5c"
      type: object
    User:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 23
        - $ref: '#/components/schemas/UserIn'
    ErrorResponse:
      properties:
        internalError:
          type: number
          format: int32
        message:
          type: string
      type: object
    ResourceNotFound:
      allOf:
        - type: object
        - $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      allOf:
        - type: object
        - $ref: '#/components/schemas/ErrorResponse'
      example:
        internalError: 999
        message: "Internal Server Error"
    BadRequest:
      allOf:
        - type: object
        - $ref: '#/components/schemas/BadRequest'
      example:
        internalError: 400
        message: "BadRequest"